API model and test improvements.
- Refined/simplified API data model and dataset resource
- Tidied up REST api test framework
- Fixed conditional request logic handling for wildcard matches
- Implemented conditional request logic in Record resource.
- Added full conditional request tests for record resource using all
conditional request header types.
- Fixed issue where createDate and modifiedDate not equal on new entity

+++ src/main/java/uk/gov/ea/datareturns/web/resource/v1/RecordResource.java
@@ -33,9 +33,8 @@ import javax.ws.rs.core.Context;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.UriBuilder;
 import javax.ws.rs.core.UriInfo;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
+import java.util.stream.Collectors;
 
 import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
 import static javax.ws.rs.core.MediaType.APPLICATION_XML;
@@ -63,6 +62,7 @@ public class RecordResource {
     @Context
     private UriInfo uriInfo;
 
+    // TODO: Graham, we are hard-coded to DataSamplePayload here, we should use a factory inside the service layer to prevent this
     private SubmissionService<DataSamplePayload, ?, ?> submissionService;
 
     /**
@@ -98,27 +98,32 @@ public class RecordResource {
     )
     @ApiResponses(value = {
             @ApiResponse(code = 200, message = "OK", response = EntityListResponse.class),
-            @ApiResponse(code = 404, message = "Not Found - The `dataset_id` parameter did not match a known resource") })
+            @ApiResponse(
+                    code = 404,
+                    message = "Not Found - The `dataset_id` parameter did not match a known resource",
+                    response = ErrorResponse.class
+            )
+    })
     public Response listRecords(
             @PathParam("dataset_id") @Pattern(regexp = "[A-Za-z0-9_-]+")
             @ApiParam("The unique identifier for the target dataset") final String datasetId)
             throws Exception {
 
         DatasetEntity datasetEntity = submissionService.getDataset(datasetId);
-
+        Response.ResponseBuilder rb;
         if (datasetEntity == null) {
-            return Response.status(Response.Status.NOT_FOUND).build();
+            rb = ErrorResponse.DATASET_NOT_FOUND.toResponseBuilder();
+        } else {
+            List<RecordEntity> records = submissionService.getRecords(datasetEntity);
+            List<EntityReference> result = new ArrayList<>();
+            for (RecordEntity record : records) {
+                UriBuilder ub = uriInfo.getAbsolutePathBuilder();
+                result.add(new EntityReference(record.getIdentifier(), ub.path(record.getIdentifier()).build().toASCIIString()));
+            }
+            EntityListResponse rw = new EntityListResponse(result);
+            rb = rw.toResponseBuilder();
         }
-
-        List<RecordEntity> records = submissionService.getRecords(datasetEntity);
-
-        List<EntityReference> result = new ArrayList<>();
-        for (RecordEntity record : records ) {
-            UriBuilder ub = uriInfo.getAbsolutePathBuilder();
-            result.add(new EntityReference(record.getIdentifier(), ub.path(record.getIdentifier()).build().toASCIIString()));
-        }
-        EntityListResponse rw = new EntityListResponse(result);
-        return rw.toResponseBuilder().build();
+        return rb.build();
     }
 
     /**
@@ -160,33 +165,89 @@ public class RecordResource {
     )
             throws Exception {
 
+        Response.ResponseBuilder rb;
         DatasetEntity datasetEntity = submissionService.getDataset(datasetId);
 
         if (datasetEntity == null) {
-            return Response.status(Response.Status.NOT_FOUND).build();
+            rb = ErrorResponse.DATASET_NOT_FOUND.toResponseBuilder();
+        } else if (batchRequest.getRequests().isEmpty()) {
+            rb = ErrorResponse.MULTISTATUS_REQUEST_EMPTY.toResponseBuilder();
+        } else {
+
+            Map<String, Response.ResponseBuilder> preconditionFailures = new HashMap<>();
+            Map<String, Response.Status> responses = new HashMap<>();
+
+            // TODO: Graham, we are hard-coded to DataSamplePayload here, we should use a factory inside the service layer to prevent this
+
+            // Build requests for the service layer
+            List<SubmissionService.ObservationIdentifierPair<DataSamplePayload>> submissible = new ArrayList<>();
+            for (BatchRecordRequestItem request : batchRequest.getRequests()) {
+                RecordEntity recordEntity = submissionService.getRecord(datasetEntity, request.getRecordId());
+
+                Response.Status defaultResponse = recordEntity != null ? Response.Status.OK : Response.Status.CREATED;
+                responses.put(request.getRecordId(), defaultResponse);
+
+                Response.ResponseBuilder tempRb = checkStoragePreconditions(recordEntity, request.getPreconditions());
+                if (tempRb == null) {
+                    submissible.add(new SubmissionService.ObservationIdentifierPair(request.getRecordId(), request.getPayload()));
+                } else {
+                    preconditionFailures.put(request.getRecordId(), tempRb);
+                }
+            }
+
+            // Create and validate records
+            Map<String, RecordEntity> recordEntities = submissionService.createRecords(datasetEntity, submissible)
+                    .stream()
+                    .collect(Collectors.toMap(RecordEntity::getIdentifier, e -> e));
+            submissionService.validate(recordEntities.values());
+
+            // Build the responses
+            MultiStatusResponse multiResponse = new MultiStatusResponse();
+            for (BatchRecordRequestItem request : batchRequest.getRequests()) {
+                UriBuilder ub = uriInfo.getAbsolutePathBuilder();
+                MultiStatusResponse.Response responseItem = new MultiStatusResponse.Response();
+                responseItem.setId(request.getRecordId());
+
+                RecordEntity recordEntity = recordEntities.get(request.getRecordId());
+                Response.ResponseBuilder tempRb = preconditionFailures.get(request.getRecordId());
+
+                if (tempRb == null) {
+                    // Preconditions passed, service the request
+                    Record record = recordAdaptor.convert(recordEntity);
+
+                    responseItem.setCode(responses.get(request.getRecordId()).getStatusCode());
+                    responseItem.setHref(ub.path(datasetId).path(request.getRecordId()).build().toASCIIString());
+                    responseItem.setEntityTag(Preconditions.createEtag(record).toString());
+                    responseItem.setLastModified(record.getLastModified());
+                } else {
+                    // Preconditions failed, build a response item from the ResponseBuilder returned by the preconditions checks
+                    Response response = tempRb.build();
+                    responseItem.setCode(response.getStatus());
+
+                    if (recordEntity != null) {
+                        // Entity does exist so populate the href
+                        responseItem.setHref(ub.path(request.getRecordId()).build().toASCIIString());
+                    }
+                }
+                multiResponse.addResponse(responseItem);
+            }
+            rb = multiResponse.toResponseBuilder();
         }
+        return rb.build();
+    }
 
-        List<SubmissionService.ObservationIdentifierPair<DataSamplePayload>> list = new ArrayList<>();
-
-        for (BatchRecordRequestItem request : batchRequest.getRequests()) {
-            list.add(new SubmissionService.ObservationIdentifierPair(request.getRecordId(), request.getPayload()));
+    private Response.ResponseBuilder checkStoragePreconditions(final RecordEntity recordEntity, Preconditions preconditions) {
+        Response.ResponseBuilder rb = null;
+        if (preconditions != null) {
+            if (recordEntity == null) {
+                rb = preconditions.evaluatePreconditions();
+            } else {
+                Record existingRecord = recordAdaptor.convert(recordEntity);
+                Date lastModified = Date.from(recordEntity.getLastChangedDate());
+                rb = preconditions.evaluatePreconditions(lastModified, Preconditions.createEtag(existingRecord));
+            }
         }
-
-        // Create and validate records
-        List<RecordEntity> recordEntities = submissionService.createRecords(datasetEntity, list);
-        submissionService.validate(recordEntities);
-
-        MultiStatusResponse multiResponse = new MultiStatusResponse();
-        for (BatchRecordRequestItem request : batchRequest.getRequests()) {
-            UriBuilder ub = uriInfo.getAbsolutePathBuilder();
-            MultiStatusResponse.Response response = new MultiStatusResponse.Response();
-            response.setId(request.getRecordId());
-            response.setCode(Response.Status.CREATED.getStatusCode());
-            response.setHref(ub.path(request.getRecordId()).build().toASCIIString());
-            multiResponse.addResponse(response);
-        }
-
-        return Response.status(207).entity(multiResponse).build();
+        return rb;
     }
 
     /**
@@ -223,23 +284,27 @@ public class RecordResource {
     )
             throws Exception {
 
+        Response.ResponseBuilder rb;
         DatasetEntity datasetEntity = submissionService.getDataset(datasetId);
 
         if (datasetEntity == null) {
-            return Response.status(Response.Status.NOT_FOUND).build();
+            rb = ErrorResponse.DATASET_NOT_FOUND.toResponseBuilder();
+        } else {
+            RecordEntity recordEntity = submissionService.getRecord(datasetEntity, recordId);
+            if (recordEntity == null) {
+                rb = ErrorResponse.RECORD_NOT_FOUND.toResponseBuilder();
+            } else {
+                Record record = recordAdaptor.convert(recordEntity);
+                rb = preconditions.evaluatePreconditions(record.getLastModified(), Preconditions.createEtag(record));
+
+                if (rb == null) {
+                    resolveLinks(datasetEntity.getIdentifier(), record);
+                    RecordEntityResponse responseWrapper = new RecordEntityResponse(Response.Status.OK, record);
+                    rb = responseWrapper.toResponseBuilder();
+                }
+            }
         }
-
-        RecordEntity recordEntity = submissionService.getRecord(datasetEntity, recordId);
-
-        if (recordEntity == null) {
-            return Response.status(Response.Status.NOT_FOUND).build();
-        }
-
-        Record record = recordAdaptor.convert(recordEntity);
-        resolveLinks(datasetEntity.getIdentifier(), record);
-
-        RecordEntityResponse responseWrapper = new RecordEntityResponse(Response.Status.OK, record);
-        return Response.status(Response.Status.OK).entity(responseWrapper).build();
+        return rb.build();
     }
 
     /**
@@ -255,45 +320,72 @@ public class RecordResource {
     @PUT
     @Path("/{record_id}")
     @ApiOperation(value = "Create or update record",
-            notes = "Create or update the record with the given `record_id` for the dataset with the given `dataset_id`.",
-            response = RecordEntityResponse.class
+            notes = "Create or update the record with the given `record_id` for the dataset with the given `dataset_id`."
     )
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "OK - an existing record was successfully updated."),
-            @ApiResponse(code = 201, message = "Created - a new record was created."),
+            @ApiResponse(code = 200, message = "OK - an existing record was successfully updated.", response = RecordEntityResponse.class),
+            @ApiResponse(code = 201, message = "Created - a new record was created.", response = RecordEntityResponse.class),
             @ApiResponse(
                     code = 404,
                     message = "Not Found - The `dataset_id` or `record_id` parameter did not match a known resource - "
-                            + "see the `meta` structure in the response envelope for more detail"
+                            + "see the `meta` structure in the response envelope for more detail",
+                    response = ErrorResponse.class
             ),
-            @ApiResponse(code = 412, message = "Precondition Failed - see conditional request documentation")
+            @ApiResponse(
+                    code = 412,
+                    message = "Precondition Failed - see conditional request documentation",
+                    response = ErrorResponse.class
+            )
     })
     public Response putRecord(
             @PathParam("dataset_id")
             @Pattern(regexp = "[A-Za-z0-9_-]+")
             @ApiParam("The unique identifier for the target dataset") final String datasetId,
+
             @PathParam("record_id")
             @Pattern(regexp = "[A-Za-z0-9_-]+")
             @ApiParam("The unique identifier for the target record") final String recordId,
             @ApiParam("The data payload to be associated with the record") final Payload payload,
+
             @BeanParam Preconditions preconditions
     )
             throws Exception {
 
+        Response.ResponseBuilder rb;
         DatasetEntity datasetEntity = submissionService.getDataset(datasetId);
 
         if (datasetEntity == null) {
-            return Response.status(Response.Status.NOT_FOUND).build();
+            rb = ErrorResponse.DATASET_NOT_FOUND.toResponseBuilder();
+        } else {
+            RecordEntity existingEntity = submissionService.getRecord(datasetEntity, recordId);
+
+            Response.Status entityResponseStatus;
+
+            if (existingEntity == null) {
+                // Create operation
+                rb = preconditions.evaluatePreconditions();
+                entityResponseStatus = Response.Status.CREATED;
+            } else {
+                // Update operation
+                Record existing = recordAdaptor.convert(existingEntity);
+                rb = preconditions
+                        .evaluatePreconditions(Date.from(existingEntity.getLastChangedDate()), Preconditions.createEtag(existing));
+                entityResponseStatus = Response.Status.OK;
+            }
+
+            if (rb == null) {
+                // Preconditions passed, create/update the record
+                existingEntity = submissionService
+                        .createRecord(datasetEntity, new SubmissionService.ObservationIdentifierPair(recordId, payload));
+                submissionService.validate(existingEntity);
+
+                Record record = recordAdaptor.convert(existingEntity);
+                resolveLinks(datasetId, record);
+                RecordEntityResponse response = new RecordEntityResponse(entityResponseStatus, record);
+                rb = response.toResponseBuilder();
+            }
         }
-
-        // Create and validate the records
-        RecordEntity recordEntity = submissionService.createRecord(datasetEntity, new SubmissionService.ObservationIdentifierPair(recordId, payload));
-        submissionService.validate(recordEntity);
-
-        Record record = recordAdaptor.convert(recordEntity);
-        resolveLinks(datasetId, record);
-        RecordEntityResponse response = new RecordEntityResponse(Response.Status.OK, record);
-        return Response.status(Response.Status.OK).entity(response).build();
+        return rb.build();
     }
 
     /**
@@ -315,9 +407,14 @@ public class RecordResource {
             @ApiResponse(
                     code = 404,
                     message = "Not Found - The `dataset_id` or `record_id` parameter did not match a known resource - "
-                            + "see the `meta` structure in the response envelope for more detail"
+                            + "see the `meta` structure in the response envelope for more detail",
+                    response = ErrorResponse.class
             ),
-            @ApiResponse(code = 412, message = "Precondition Failed - see conditional request documentation")
+            @ApiResponse(
+                    code = 412,
+                    message = "Precondition Failed - see conditional request documentation",
+                    response = ErrorResponse.class
+            )
     })
     public Response deleteRecord(
             @PathParam("dataset_id")
@@ -330,21 +427,27 @@ public class RecordResource {
     )
             throws Exception {
 
+        Response.ResponseBuilder rb;
         DatasetEntity datasetEntity = submissionService.getDataset(datasetId);
 
         if (datasetEntity == null) {
-            return Response.status(Response.Status.NOT_FOUND).build();
+            rb = ErrorResponse.DATASET_NOT_FOUND.toResponseBuilder();
+        } else {
+            RecordEntity recordEntity = submissionService.getRecord(datasetEntity, recordId);
+            if (recordEntity == null) {
+                rb = ErrorResponse.RECORD_NOT_FOUND.toResponseBuilder();
+            } else {
+                Record record = recordAdaptor.convert(recordEntity);
+                rb = preconditions.evaluatePreconditions(record.getLastModified(), Preconditions.createEtag(record));
+
+                if (rb == null) {
+                    // Preconditions passed, delete the record
+                    submissionService.removeRecord(recordEntity);
+                    rb = Response.status(Response.Status.NO_CONTENT);
+                }
+            }
         }
-
-        RecordEntity recordEntity = submissionService.getRecord(datasetEntity, recordId);
-
-        if (recordEntity == null) {
-            return Response.status(Response.Status.NOT_FOUND).build();
-        }
-
-        submissionService.removeRecord(recordEntity);
-
-        return Response.status(Response.Status.NO_CONTENT).build();
+        return rb.build();
     }
 
     private void resolveLinks(String datasetId, Record record) {
